Roll analyzer to 0.27.2-alpha.1
R=jmesserly@google.com

Review URL: https://codereview.chromium.org/1641643004 .

+++ pkg/dev_compiler/tool/test.sh
@@ -15,7 +15,9 @@ cd $( dirname "${BASH_SOURCE[0]}" )/..
 
 # Delete codegen expectation files to be sure that if a test fails to compile
 # we don't erroneously pick up the old version.
-rm -r test/codegen/expect || fail
+if [ -d test/codegen/expect ]; then
+  rm -r test/codegen/expect || fail
+fi
 
 # Make sure we don't run tests in code coverage mode.
 # this will cause us to generate files that are not part of the baseline
